requirements.txt
Flask==1.1.1
flask-restful==0.3.6
psycopg2-binary==2.8.4